+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
+* Added data file, configuration file and configuration parsing.
 * Added argument parsing for file path.
 * Added pytest-mock.
 * Added pytest.